% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_between.R
\name{beds_available_between}
\alias{beds_available_between}
\title{beds_available_between}
\value{
\code{data.frame/logical} after filtering/applying conditional to the appropriate columns
}
\description{
Context sensitive quick filtering or output of logical based on \code{start} and \code{end} Dates.
}
\details{
Context-sensitive: Automatically detects if nested inside of \link[dplyr]{filter} call, if so returns \code{logical} instead of \code{data.frame}
}
\examples{
\dontrun{
library(dplyr)
ReportStart = Sys.Date() - lubridate::weeks(4)
ReportEnd = Sys.Date()
identical(
served_between(qpr_leavers),
qpr_leavers \%>\% filter(served_between(.))
)
}
}
\seealso{
Other _between: 
\code{\link{between_df}},
\code{\link{entered_between}},
\code{\link{exited_between}},
\code{\link{operating_between}},
\code{\link{served_between}},
\code{\link{stayed_between}}
}
\concept{_between}
