% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{between_df}
\alias{between_df}
\title{between_df}
\usage{
between_df(., status, start = ReportStart, end = ReportEnd, lgl = FALSE)
}
\arguments{
\item{.}{\code{(data.frame/tibble)} Input to be filtered. In a \code{magrittr}
pipe this will always be the first object}

\item{status}{\code{(character)} One of:
\itemize{
\item{\code{"served"/"se"}}{ Equivalent of \code{served_between}}
\item{\code{"stayed"/"st"}}{ Equivalent of \code{stayed_between}}
\item{\code{"entered"/"en"}}{ Equivalent of \code{entered_between}}
\item{\code{"exited"/"ex"}}{ Equivalent of \code{exited_between}}
\item{\code{"operating"/"op"}}{ Equivalent of \code{operating_between}}
\item{\code{"beds_available"/"be"/"ba"}}{ Equivalent of \code{beds_available_between}}
}
that specifies the type of function to be performed}

\item{start}{\code{character/Date} of the end of the timeframe. Characters
in format mdY, Ymd, dmY acceptable.  \strong{Default \code{ReportStart}} will be
automatically retrieved from parent environments if not specified. If end is
named other than ReportStart, it must be specified.}

\item{end}{\code{character/Date} of the end of the timeframe. Characters in
format mdY, Ymd, dmY acceptable.  \strong{Default \code{ReportEnd}} will be
automatically retrieved from parent environments if not specified. If end is
named other than ReportEnd, it must be specified.}

\item{lgl}{\code{logical} Flag to force logical vector output. \strong{Default \code{FALSE}}}
}
\value{
\code{data.frame/logical} after filtering/applying conditional to the
appropriate columns
}
\description{
Context sensitive quick filtering or output of logical based on
\code{start} and \code{end} Dates.
}
\details{
Context-sensitive: Automatically detects if nested inside
of \link[dplyr]{filter} call, if so returns \code{logical} instead of \code{data.frame}
}
\examples{
\dontrun{
library(dplyr)
ReportStart = Sys.Date() - lubridate::weeks(4)
ReportEnd = Sys.Date()
identical(
served_between(qpr_leavers),
qpr_leavers \%>\% filter(served_between(.))
)
}
}
\seealso{
Other _between: 
\code{\link{beds_available_between}()},
\code{\link{entered_between}()},
\code{\link{exited_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{_between}
\keyword{Internal}
