% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{check_dates}
\alias{check_dates}
\title{check_dates}
\usage{
check_dates(start, end)
}
\arguments{
\item{start}{Default = ReportStart, character/date of the start of the date
range. Characters in format mdY, Ymd, dmY acceptable. Will be automatically
retrieved from parent environments if not specified. If start is named other
than ReportStart, it must be specified.}

\item{end}{Default = ReportEnd, character/date of the end of the date range.
Characters in format mdY, Ymd, dmY acceptable. Will be automatically
retrieved from parent environments if not specified. If end is named other
than ReportEnd, it must be specified.}
}
\description{
coerce start/end input values to dates
}
\keyword{Internal}
