% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{check_names}
\alias{check_names}
\title{check_names}
\usage{
check_names(., nms)
}
\arguments{
\item{.}{\code{(data.frame/tibble)} data input}

\item{nms}{\code{(character)} Vector of column names that must be present}
}
\description{
Ensure the appropriate names required for filtering are present in the data and if not, give informative error.
}
\keyword{Internal}
