% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{enhanced_yes_no_translator}
\alias{enhanced_yes_no_translator}
\title{Translate HUD reference numbers \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
enhanced_yes_no_translator(ReferenceNo)
}
\arguments{
\item{ReferenceNo}{\code{(integer)} vector. Quasiquoted column name if used in \link[dplyr]{mutate}.}
}
\value{
\code{(character)} Human-readable corresponding character vector.
}
\description{
Deprecated in favor of \code{hud_translations}. Translate HUD reference numbers
}
