% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{exited_between}
\alias{exited_between}
\title{exited_between}
\usage{
exited_between(., start = ReportStart, end = ReportEnd, lgl = FALSE)
}
\arguments{
\item{.}{data.frame/tibble. In a \code{magrittr}, pipe this will always be the
first object}

\item{start}{Default = ReportStart, character/date of the start of the date
range. Characters in format mdY, Ymd, dmY acceptable. Will be automatically
retrieved from parent environments if not specified. If start is named other
than ReportStart, it must be specified.}

\item{end}{Default = ReportEnd, character/date of the end of the date range.
Characters in format mdY, Ymd, dmY acceptable. Will be automatically
retrieved from parent environments if not specified. If end is named other
than ReportEnd, it must be specified.}

\item{lgl}{Default = FALSE, logical, flag to force logical vector output.}
}
\value{
\code{data.frame/logical} after filtering/applying conditional to the
appropriate columns
}
\description{
Filters a dataframe of Enrollments where the Exit Date falls
within the date range specified. Used to report on households leaving a
project. Useful for measuring outcomes.
}
\details{
Context-sensitive: Automatically detects if nested inside
of \link[dplyr]{filter} call, if so returns \code{logical} instead of \code{data.frame}.
Enrollment-related functions require Enrollment data elements such as Entry
Date, Exit Date, Move In Date, and Project Type (of the Enrollment's Project).
Project Descriptor-related functions require Project data elements such as
Operating Start/End dates and Bed Inventory Start/End dates.
}
\examples{
\dontrun{
library(dplyr)
ReportStart <- Sys.Date() - lubridate::weeks(4)
ReportEnd <- Sys.Date()
identical(
served_between(qpr_leavers),
qpr_leavers \%>\% filter(served_between(.))
)
}

}
\seealso{
Other _between: 
\code{\link{beds_available_between}()},
\code{\link{between_df}()},
\code{\link{entered_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{_between}
