% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{hud_translate}
\alias{hud_translate}
\title{Translate Numeric/Character from a Hash Table with columns Value/Text}
\usage{
hud_translate(x, hash)
}
\arguments{
\item{x}{\code{(character/numeric)} vector to translate}

\item{hash}{\code{(data.frame)} with Value/Text numeric/character corresponding to the value to translate}
}
\value{
\code{(numeric/character)} Whichever class is opposite the input vector
}
\description{
Translate Numeric/Character from a Hash Table with columns Value/Text
}
