% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\docType{data}
\name{hud_translations}
\alias{hud_translations}
\title{Translate HUD Coding of Data Elements}
\format{
An object of class \code{list} of length 88.
}
\usage{
hud_translations
}
\arguments{
\item{.x}{\code{(character/numeric)} The values to translate}
}
\value{
\code{(character/numeric)} equivalent, depending on the input
}
\description{
Translate values from HUD Data elements between values or text
}
\keyword{datasets}
