% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{living_situation}
\alias{living_situation}
\title{Living Situation Reference Number Translation\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
living_situation(ReferenceNo)
}
\arguments{
\item{ReferenceNo}{\code{(integer)} Reference number for living situation type}
}
\value{
\code{(character)} Human-readable living situation type
}
\description{
Deprecated in favor of \code{hud_translations}.  Return a human-readable living situation character vector provided with an integer reference number
}
