% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{make_date.character}
\alias{make_date.character}
\title{Coerce various inputs to \code{(Date)}}
\usage{
\method{make_date}{character}(x)
}
\arguments{
\item{x}{\code{(character/numeric/POSIXct/lt)}}
}
\value{
\code{(Date)}
}
\description{
Coerce various inputs to \code{(Date)}
}
