% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{project_type}
\alias{project_type}
\title{Project Type Reference Number Translation}
\usage{
project_type(ReferenceNo)
}
\arguments{
\item{ReferenceNo}{\code{(integer)} Reference number for project type}
}
\value{
\code{(character)} Human-readable project type
}
\description{
Return a human-readable project type provided a reference number
}
