% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{replace_yes_no}
\alias{replace_yes_no}
\title{Yes/No to numeric binary \code{1/0}}
\usage{
replace_yes_no(column_name)
}
\arguments{
\item{column_name}{\code{(character)} Column vector. Quasiquoted name if used in \link[dplyr]{mutate}.}
}
\value{
\code{(Integer)} Binary encoded integer vector
}
\description{
Change a binary character \code{"No"} or \code{NA} to \code{0} and everything else to \code{1}
}
