% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{beds_available_between}
\alias{beds_available_between}
\title{beds_available_between}
\usage{
beds_available_between(., start = ReportStart, end = ReportEnd, lgl = FALSE)
}
\arguments{
\item{.}{\code{(data.frame/tibble)} Input to be filtered. In a \code{magrittr}
pipe this will always be the first object}

\item{start}{\code{character/Date} of the end of the timeframe. Characters
in format mdY, Ymd, dmY acceptable.  \strong{Default \code{ReportStart}} will be
automatically retrieved from parent environments if not specified. If end is
named other than ReportStart, it must be specified.}

\item{end}{\code{character/Date} of the end of the timeframe. Characters in
format mdY, Ymd, dmY acceptable.  \strong{Default \code{ReportEnd}} will be
automatically retrieved from parent environments if not specified. If end is
named other than ReportEnd, it must be specified.}

\item{lgl}{\code{logical} Flag to force logical vector output. \strong{Default \code{FALSE}}}
}
\value{
\code{data.frame/logical} after filtering/applying conditional to the
appropriate columns
}
\description{
Context sensitive quick filtering or output of logical based on
\code{start} and \code{end} Dates.
}
\details{
Context-sensitive: Automatically detects if nested inside
of \link[dplyr]{filter} call, if so returns \code{logical} instead of \code{data.frame}
}
\examples{
\dontrun{
library(dplyr)
ReportStart = Sys.Date() - lubridate::weeks(4)
ReportEnd = Sys.Date()
identical(
served_between(qpr_leavers),
qpr_leavers \%>\% filter(served_between(.))
)
}
}
\seealso{
Other _between: 
\code{\link{between_df}()},
\code{\link{entered_between}()},
\code{\link{exited_between}()},
\code{\link{operating_between}()},
\code{\link{served_between}()},
\code{\link{stayed_between}()}
}
\concept{_between}
