% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{enhanced_yes_no_translator}
\alias{enhanced_yes_no_translator}
\title{Translate HUD reference numbers}
\usage{
enhanced_yes_no_translator(ReferenceNo)
}
\arguments{
\item{ReferenceNo}{\code{(integer)} vector. Quasiquoted column name if used in \link[dplyr]{mutate}.}
}
\value{
\code{(character)} Human-readable corresponding character vector.
}
\description{
Translate HUD reference numbers
}
