% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{project_type}
\alias{project_type}
\title{Project Type Reference Number Translation \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
project_type(ReferenceNo)
}
\arguments{
\item{ReferenceNo}{\code{(integer)} Reference number for project type}
}
\value{
\code{(character)} Human-readable project type
}
\description{
Deprecated in favor of \code{hud_translations}. Return a human-readable project type provided a reference number
}
