% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_reference.R
\name{translate_HUD_yes_no}
\alias{translate_HUD_yes_no}
\title{Translate HUD reference numbers}
\usage{
translate_HUD_yes_no(column_name)
}
\arguments{
\item{column_name}{\code{(integer)} Column vector. Quasiquoted column name if used in \link[dplyr]{mutate}.}
}
\value{
\code{(character)} Human-readable corresponding character vector.
}
\description{
This function translates the HUD csv 1.7 and 1.8 lists and returns yes, no, or unknown as appropriate
}
